[feat/#176] 리턴 타입 추가

+++ presentation/consumer-payment/src/main/java/com/ssing/presentation/consumerpayment/navigation/PaymentNavigation.kt
@@ -14,7 +14,7 @@ fun NavController.navigateToConsumerPayment(matchingRequestId: Long) =
     navigate(ConsumerPayment(matchingRequestId))
 
 fun NavGraphBuilder.consumerPaymentNavGraph(
-    navigateToLesson: () -> Unit,
+    navigateToLesson: (Long) -> Unit,
     navigateToHome: () -> Unit,
 ) {
     slideComposable<ConsumerPayment> {
